omg!  removed lucid project (the managed class library) and recreated it as facade in an attempt to fix broken project dependencies (thought having two things named lucid was confusing VS).  in so doing ran into issues with 32 vs 64 bit targets (since had to change c++/cli stuff).  so, initial changes to make sure 32 and 64 bit targets work.  going to have to clean up the file reading code more but they work for now.

+++ main/lucid/gigl/Batched.h
@@ -124,7 +124,7 @@ namespace gigl {
 			Batch(size_t maximum)
 				: maximum(maximum)
 			{
-				batch.reset(gal::VertexBuffer::create(gal::VertexBuffer::USAGE_DYNAMIC, maximum, sizeof(I)));
+				batch.reset(gal::VertexBuffer::create(gal::VertexBuffer::USAGE_DYNAMIC, int32_t(maximum), sizeof(I)));
 			}
 
 			virtual ~Batch() = default;
@@ -160,7 +160,7 @@ namespace gigl {
 					batch->unlock();
 
 					pipeline.setVertexStream(1, batch.get());
-					mesh->drawInstanced(count);
+					mesh->drawInstanced(int32_t(count));
 
 					index += count;
 				}
